% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxlongrange.R
\name{jump.scan}
\alias{jump.scan}
\title{Scan over predicted coordinates for areas of long-range jumps.}
\usage{
jump.scan(dates, coord.longlat, newcoord.longlat = NULL, r = NULL, params,
  Albers = NULL, covar = NULL)
}
\arguments{
\item{dates}{vector of length \eqn{N} representing the observed time when the process first appears.}

\item{coord.longlat}{matrix of dimension \eqn{N} x 2 representing the longitude and lattitude of each observation.}

\item{newcoord.longlat}{matrix of dimension \eqn{n} x 2 representing the longitude and lattitude locations to estimate the jumps. The default is \code{coord.longlat}.}

\item{r}{radius about the jumps. The default is the 2.5 percent quantile of the distance matrix of coordinates.}

\item{params}{estiamted parameter \eqn{\Theta = (\beta^\prime, \eta)^\prime}{\Theta = (\beta', \eta)'}  for the Gaussian process model using \code{\link{localgrad}}. \eqn{\beta} represents the regression coefficients and \eqn{\eta} represents the covariance parameters.}

\item{Albers}{vector of length \eqn{2} representing the parameters of Albers equal area conic projection. If process of interest is on the longitude and lattitude domain, user must specify \code{Albers}. If specified, use the Albers equal area conic projection. The default setting does not use projections. See \code{\link{mapproject}} for an explanation.}

\item{covar}{matrix of dimension \eqn{N} x \eqn{p} of explanatory variables except for coordinates.}
}
\value{
\code{jump.scan} returns an object of class \dQuote{\code{jump.scan}}, which is a list containing the following components.
\item{coord.longlat}{ matrix of dimension \eqn{N} x 2 representing the longitude and lattitude of each observation. }
\item{newcoord.longlat}{ matrix of dimension \eqn{n} x 2 representing longitude and lattitude locations to estimate the jumps.}
\item{conv.factor}{ conversion factor to minimize distortion between Albers conic projected coordinates and longitude latitude coordinates. If projection is not used, the \code{conv.factor} is 1.}
\item{r}{ radius about jumps.}
\item{all.sides}{ Matrix of dimension \eqn{n} x 4 representing significant sides. Return whether each side has a significantly negative gradient (-1), significantly positive gradient (+1) or non-significant gradient (0). }
}
\description{
Scan over predicted coordinates for areas of long-range jumps.
}
\details{
This function directly tests if there is a significant radial expansion around a predicted point as in Goldstein et al. (2018), using the distribution of the gradient around a point (Banerjee and Gelfand, 2006).
When centered on each location, we can test the gradient normal to four sides of a box with sides of radius r.
As a heuristic we say if the spread is significantly outside of at least two sides of the box, and does not go significantly into any side of the box,
we can flag the location as a potential site of a long-range jump.
}
\examples{
\dontrun{

###########    hemlock data example    ##########
data(hemlock)
coord = cbind(hemlock$long,hemlock$lat)    # longitude latitude coordinates
dates = hemlock$first.year                 # quaratine data by county

# adaptive mcmc algorithm
# specify batches, batch length, target acceptance rate
amcmc=list(n.batch=10,batch.length=100,accept.rate=0.3)
out.grad = localgrad(dates=dates,coord.longlat=coord, Albers=c(29.5,45.5),
n.samp=1000,thin=10, amcmc=amcmc)

# posterior mean of parameters
par.mean = apply(out.grad$post.samp,2,mean)

# test for long range jump
out.jump = jump.scan(dates=dates, coord.longlat=coord, params = par.mean, Albers= c(29.5,45.5) )
}
}
\references{
Banerjee, S., and Gelfand, A. E. (2006). Bayesian wombling: Curvilinear gradient assessment under spatial process models. \emph{Journal of the American Statistical Association}, \bold{101}(476), 1487-1501.

Goldstein, J., Jaewoo Park, Haran, M., Liebhold, A., and Bjornstad, O. N. (2018). Quantifying Spatio-Temporal Variation of Invasion Spread.\emph{arXiv preprint arXiv:1506.02685v3}
}
\seealso{
\code{\link{plotboxes}}
}
