% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localgrad.R
\name{localgrad}
\alias{localgrad}
\title{Estimate local gradients.}
\usage{
localgrad(dates, covar = NULL, coord.longlat, newcoord.longlat = NULL,
  Albers = NULL, n.samp = 10000, thin = 10, knots = NA, amcmc = NA,
  verbose = FALSE)
}
\arguments{
\item{dates}{vector of length \eqn{N} representing the observed time when the process first appears.}

\item{covar}{matrix of dimension \eqn{N} x \eqn{p} representing explanatory variables except for coordinates. If not specified, only \code{coord.longlat} is used as an explanatroy variable.}

\item{coord.longlat}{matrix of dimension \eqn{N} x 2 representing the longitude and lattitude of each observation.}

\item{newcoord.longlat}{matrix of dimension \eqn{n} x 2 representing the longitude and lattitude locations to estimate the gradient. The default is \code{coord.longlat}.}

\item{Albers}{vector of length \eqn{2} representing the parameters of Albers equal area conic projection. If process of interest is on the longitude and lattitude domain, user must specify \code{Albers}. If not specified, projection is not conducted and mean speed of spread is not evaluated in the \code{\link{summary.localgrad}}. See \code{\link{mapproject}} for an explanation.}

\item{n.samp}{total number of posterior samples. Defaults to 10,000.}

\item{thin}{the thinning interval between consecutive observations. Defaults to 10.}

\item{knots}{if specified, use kernel convolution method over knot process to speed up computation. See \code{\link{spLM}} for an explanation.}

\item{amcmc}{if specified, use adaptive mcmc algorithm. See \code{\link{spLM}} for an explanation.}

\item{verbose}{if TRUE, progress of the sampler is printed to the screen. Otherwise, nothing is printed.}
}
\value{
\code{localgrad} returns an object of class \dQuote{\code{localgrad}}, which is a list containing the following components.
\item{dates}{ vector of length \eqn{N} representing the observed time when the process first appears. }
\item{coord.longlat}{ matrix of dimension \eqn{N} x 2 representing the longitude and lattitude of each observation. }
\item{newcoord.longlat}{ matrix of dimension \eqn{n} x 2 representing locations to estimate the gradient. }
\item{coord }{ matrix of dimension \eqn{N} x 2 representing the Albers conic projected observation. If projection is not used, \code{coord} is identical to \code{coord.longlat}. }
\item{newcoord }{ matrix of dimension \eqn{n} x 2 representing the Albers conic projected locations to be estimated. If projection is not used, \code{newcoord} is identical to \code{newcoord.longlat}. }
\item{conv.factor}{ conversion factor to minimize distortion between Albers conic projected coordinates and longitude latitude coordinates. If projection is not used, \code{conv.factor} is 1.}
\item{samples }{ list of posterior samples for the gradient at each location in newcoord. Each posterior sample consists of an \eqn{x} and \eqn{y} component of the gradient.}
\item{means}{ mean speed of spread at each location in newcoord. }
\item{sig}{ vector of representing whether the gradients are significant or not in the \code{newcoord}. }
\item{post.samp}{ thined posterior samples for \eqn{\Theta} from \code{\link{spLM}} function. }
}
\description{
Estimate local gradients.
}
\details{
This function estimates gradients of surface by specifying a Gaussian process model as in Goldstein et al. (2018) using theoretical results in Banerjee et al. (2003); Banerjee and Gelfand (2006).
If the surface is the waiting time to first appearance of the process, then the reciprocal of the gradient length is a measure of the invasion speed: Fast spread leads to shallow waiting time surfaces, while slow spread results in steep surfaces.

Consider \eqn{N} observations of the year of first appearance \eqn{Y(s)} at location \eqn{s}. \eqn{Y(s)} can be modeled using an isotropic
Gaussian process with mean \eqn{\mu(s)} and covariance \eqn{K(D)}, where \eqn{D} is the \eqn{N} x \eqn{N} matrix of pairwise Euclidian distances of \eqn{s}.
Then the gradient of \eqn{Y(s)} at \eqn{s}, the vector of directional derivatives can be defined as \eqn{dY(s)}.
The conditional distribution of the gradient can be derived as:

\deqn{dY(s)\mid Y,\Theta \sim N(d\mu(s)- dK(D)^\prime K(D)^{-1}(Y-\mu), -H(D)-dK(D)^\prime K(D)^{-1}dK(D) ),}{dY(s)|Y,\Theta ~ N(d\mu(s)- dK(D)'K(D)^{-1}(Y-\mu), -H(D)-dK(D)'K(D)^{-1}dK(D) ),}

where \eqn{\Theta = (\beta^\prime, \eta)^\prime}{\Theta = (\beta', \eta)'} is the parameters for the Gaussian process;
\eqn{\beta} is the regression parameter for \eqn{\mu(s)} and \eqn{\eta} is the covariance parameter in \eqn{K(D)} and \eqn{H(D)} is the Hessian matrix of \eqn{K(D)}.

\code{localgrad} function conducts Bayesian inference for \eqn{\Theta} through the \code{spLM} function.
Simulation from the predictive distribution of the gradient is done by composition; given each posterior sample of
\eqn{\Theta}, a sample for \eqn{dY(s)} is drawn from the conditional distribution of the gradient.
Then speed of spread and the dominant direction of the spread are calculated and tested.
}
\examples{
\dontrun{

###########    hemlock data example    ##########
data(hemlock)
coord = cbind(hemlock$long,hemlock$lat)    # longitude latitude coordinates
dates = hemlock$first.year                 # quaratine data by county

# adaptive mcmc algorithm
# specify batches, batch length, target acceptance rate
amcmc=list(n.batch=10,batch.length=100,accept.rate=0.3)
out.grad = localgrad(dates=dates,coord.longlat=coord, Albers=c(29.5,45.5),
n.samp=1000,thin=10, amcmc=amcmc)
}
}
\references{
Banerjee, S., and Gelfand, A. E. (2006). Bayesian wombling: Curvilinear gradient assessment under spatial process models. \emph{Journal of the American Statistical Association}, \bold{101}(476), 1487-1501.

Banerjee, S., Gelfand, A. E., and Sirmans, C. F. (2003). Directional rates of change under spatial process models. \emph{Journal of the American Statistical Association}, \bold{98}(464), 946--954.

Goldstein, J., Jaewoo Park, Haran, M., Liebhold, A., and Bjornstad, O. N. (2018). Quantifying Spatio-Temporal Variation of Invasion Spread.\emph{arXiv preprint arXiv:1506.02685v3}
}
\seealso{
\code{\link{spLM}}
}
