% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localgrad.R
\name{longrange.localgrad}
\alias{longrange.localgrad}
\title{Test class \dQuote{\code{localgrad}} for potential regions of long-range jumps
using a Rayleigh test of a uniform circular distribution (Jammalamadaka and Sengupta, 2001).}
\usage{
longrange.localgrad(object, r = NULL, plot = TRUE, add = FALSE,
  main = "Test for long range jumps", xlab = "Longitude",
  ylab = "Latitude", col = 1, cex = 5, ...)
}
\arguments{
\item{object}{an object of class \code{localgrad}, typically the result of a call to \code{\link{localgrad}}.}

\item{r}{radius of jump to perform the Rayleigh test. The default is the 2.5 percent quantile of the distance matrix of Albers conic projected coordinates. If projection is not used in \code{\link{localgrad}}, the default is the 2.5 percent quantile of the distance matrix of original coordinates.}

\item{plot}{TRUE: plot points around which the spread is radially outward.}

\item{add}{whether we add to the original plot or not.}

\item{main}{an overall title for the plot: see \code{\link{plot}}. The default title is "Test for long-range jumps".}

\item{xlab}{a title for the \eqn{x} axis: see \code{\link{plot}}. The default title is "Longitude".}

\item{ylab}{a title for the \eqn{y} axis: see \code{\link{plot}}. The default title is "Latitude".}

\item{col}{color code or name, see \code{\link{par}}.}

\item{cex}{character (or symbol) expansion: a numerical vector. This works as a multiple of \code{\link{par}}("cex").}

\item{\dots}{additional arguments.}
}
\value{
\code{longrange.localgrad} returns an object of class \dQuote{\code{longrange.localgrad}}, which is a list containing the following components.
\item{rayleigh.n}{ number of points in the neighborhood used for the Rayleigh test at each location. }
\item{rayleigh.z}{ summary statistic for the Rayleigh test at each location. }
\item{sig}{ binary vector, 1 for locations where the Rayleigh test is significant. }
}
\description{
Test class \dQuote{\code{localgrad}} for potential regions of long-range jumps
using a Rayleigh test of a uniform circular distribution (Jammalamadaka and Sengupta, 2001).
}
\details{
This function conducts the Rayleigh test to the gradients from the class \dQuote{\code{localgrad}}.
Then flag at the potential sites of long range jumps.
}
\examples{
\dontrun{

###########    hemlock data example    ##########
data(hemlock)
coord = cbind(hemlock$long,hemlock$lat)    # longitude latitude coordinates
dates = hemlock$first.year                 # quaratine data by county

# adaptive mcmc algorithm
# specify batches, batch length, target acceptance rate
amcmc=list(n.batch=10,batch.length=100,accept.rate=0.3)
out.grad = localgrad(dates=dates,coord.longlat=coord, Albers=c(29.5,45.5),
n.samp=1000,thin=10, amcmc=amcmc)

# gradient plot
plotgrad(out.grad,cex=1,pch=".",database="state")

# rayleigh test
out.longrange = longrange.localgrad(out.grad,add=TRUE)
}
}
\references{
Goldstein, J., Jaewoo Park, Haran, M., Liebhold, A., and Bjornstad, O. N. (2018). Quantifying Spatio-Temporal Variation of Invasion Spread.\emph{arXiv preprint arXiv:1506.02685v3}

Jammalamadaka, S. R. and Sengupta, A. (2001). Topics in circular statistics. \emph{(volume 5). World Scientific.}
}
\seealso{
\code{\link{localgrad}}
}
