% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxlongrange.R
\name{plotboxes}
\alias{plotboxes}
\title{Plot the significant long range jumps from the class \dQuote{\code{jump.scan}}.}
\usage{
plotboxes(r = NULL, object, num.sides, method = c("box", "arrow"),
  point = FALSE, ...)
}
\arguments{
\item{r}{length of boxes or arrows on the plot. The default is 1.}

\item{object}{an object of class \code{jump.scan}, typically the result of a call to \code{\link{jump.scan}}.}

\item{num.sides}{number of significant sides in the box. It should be the integer between 1 and 4.}

\item{method}{the method to be used. If method is \dQuote{\code{box}}, draw a box for significant sides. If the method is \dQuote{\code{arrow}}, draw arrows for significant sides.}

\item{point}{if TRUE, plot points where long-range jumps are detected.}

\item{\dots}{additional arguments.}
}
\description{
Plot the significant long range jumps from the class \dQuote{\code{jump.scan}}.
}
\details{
This function draws boxes representing long range jumps on the plot using an object of class \code{jump.scan}.
}
\examples{
\dontrun{

###########    hemlock data example    ##########
data(hemlock)
coord = cbind(hemlock$long,hemlock$lat)    # longitude latitude coordinates
dates = hemlock$first.year                 # quaratine data by county

# adaptive mcmc algorithm
# specify batches, batch length, target acceptance rate
amcmc=list(n.batch=10,batch.length=100,accept.rate=0.3)
out.grad = localgrad(dates=dates,coord.longlat=coord, Albers=c(29.5,45.5),
n.samp=1000,thin=10, amcmc=amcmc)

# posterior mean of parameters
par.mean = apply(out.grad$post.samp,2,mean)

# test for long range jump
out.jump = jump.scan(dates=dates, coord.longlat=coord, params = par.mean, Albers= c(29.5,45.5) )

# gradient plot
plotgrad(out.grad,cex=1,pch=".",database="state")

# long-range jump plot
plotboxes(object = out.jump, num.sides = 2, method="box")
}
}
\seealso{
\code{\link{jump.scan}}
}
