% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localgrad.R
\name{plotgrad}
\alias{plotgrad}
\title{Draw the vector field plot for the significant gradient from the class \dQuote{\code{localgrad}}.}
\usage{
plotgrad(object, main = "Speed of spread", scale = 1, database = NULL,
  xlab = expression(paste(degree, "Longitude")),
  ylab = expression(paste(degree, "Latitude")), xlim = NA, ylim = NA, ...)
}
\arguments{
\item{object}{an object of class \code{localgrad}, typically the result of a call to \code{\link{localgrad}}.}

\item{main}{an overall title for the plot: see \code{\link{plot}}. The default title is "Speed of spread".}

\item{scale}{scale of arrow vector. The default is 1.}

\item{database}{character string naming a geographical database in \code{\link{map}} function. If specified, draw geographical map on the gradient plot. The default is not using \code{\link{map}} function.}

\item{xlab}{a title for the \eqn{x} axis: see \code{\link{plot}}. The default title is "Longitude".}

\item{ylab}{a title for the \eqn{y} axis: see \code{\link{plot}}. The default title is "Latitude".}

\item{xlim}{numeric vectors of length 2, giving the ranges for \eqn{x} axis: see \code{\link{plot}}. The default is range of \eqn{x} axis in the predicted coordinates.}

\item{ylim}{numeric vectors of length 2, giving the ranges for \eqn{y} axis: see \code{\link{plot}}. The default is range of \eqn{y} axis in the predicted coordinates.}

\item{\dots}{additional arguments.}
}
\description{
Draw the vector field plot for the significant gradient from the class \dQuote{\code{localgrad}}.
}
\details{
This function draws the vector field plot for class \dQuote{\code{localgrad}}. Arrows in the plot indicate the direction and speed of spread.
As the speed of spread increases, the length of the arrows become longer.
The color of each arrow represents the first appearance time. Red indicates the earliest appearance, and blue represents the latest appearance.
}
\examples{
\dontrun{

###########    hemlock data example    ##########
data(hemlock)
coord = cbind(hemlock$long,hemlock$lat)    # longitude latitude coordinates
dates = hemlock$first.year                 # quaratine data by county

# adaptive mcmc algorithm
# specify batches, batch length, target acceptance rate
amcmc=list(n.batch=10,batch.length=100,accept.rate=0.3)
out.grad = localgrad(dates=dates,coord.longlat=coord, Albers=c(29.5,45.5),
n.samp=1000,thin=10, amcmc=amcmc)

# gradient plot
plotgrad(out.grad,cex=1,pch=".",database="state")
}
}
\seealso{
\code{\link{localgrad}}, \code{\link{map}}
}
