% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localgrad.R
\name{summary.localgrad}
\alias{summary.localgrad}
\title{Print a summary of a class \dQuote{\code{localgrad}}.}
\usage{
\method{summary}{localgrad}(object, ...)
}
\arguments{
\item{object}{an object of class \code{localgrad}, typically the result of a call to \code{\link{localgrad}}.}

\item{\dots}{additional arguments.}
}
\description{
Print a summary of a class \dQuote{\code{localgrad}}.
}
\details{
This is the summary function for class \dQuote{\code{localgrad}}. Among the predicted coordinates this function displays the number of significant locations.
If the process of interest is on the longitude and lattitude domian, this also returns the estimated mean and median speed.
}
\examples{
\dontrun{

###########    hemlock data example    ##########
data(hemlock)
coord = cbind(hemlock$long,hemlock$lat)    # longitude latitude coordinates
dates = hemlock$first.year                 # quaratine data by county

# adaptive mcmc algorithm
# specify batches, batch length, target acceptance rate
amcmc=list(n.batch=10,batch.length=100,accept.rate=0.3)
out.grad = localgrad(dates=dates,coord.longlat=coord, Albers=c(29.5,45.5),
n.samp=1000,thin=10, amcmc=amcmc)

# summary of inference results
summary(out.grad)
}
}
\seealso{
\code{\link{localgrad}}
}
